move code as per review
